restructured sourcecode in order to support more task types
