+++ src/commands/index.ts
@@ -11,5 +11,5 @@ export * from './update-organization';
 export * from './update-stacks';
 export * from './validate-stacks';
 export * from './validate-tasks';
-export * from './update-sls';
+export * from './serverless/update-sls';
 export * from './cleanup';
